feat: migrate to Log4j2 and improve parsing error logging
- Switched from Logback to Log4j2 to ensure full stack traces in logs
- Added exclusions for spring-boot-starter-logging in all dependencies
- Configured log4j2.xml with explicit exception layout
- Enhanced JacksonEventParser with detailed line-by-line error logging
- Refactored EventLogStatus to use Byte for database compatibility

+++ src/main/java/com/spark/insight/model/EventLogStatus.java
@@ -5,16 +5,16 @@ import lombok.Getter;
 
 @Getter
 public enum EventLogStatus {
-    PROCESSING(0, "Parsing"),
-    IMPORTING(1, "Importing"),
-    SUCCESS(2, "Success"),
-    FAILED(3, "Failed");
+    PROCESSING((byte) 0L, "Parsing"),
+    IMPORTING((byte) 1, "Importing"),
+    SUCCESS((byte) 2, "Success"),
+    FAILED((byte) 3, "Failed");
 
     @EnumValue
-    private final int code;
+    private final Byte code;
     private final String description;
 
-    EventLogStatus(int code, String description) {
+    EventLogStatus(Byte code, String description) {
         this.code = code;
         this.description = description;
     }
